Add better implementation for generic columns
